+++ src/modules/password-reset/domain/interfaces/password-reset-form.interface.ts
@@ -1,0 +1,4 @@
+export interface PasswordResetFormInterface {
+  newPassword: string
+  passwordConfirm: string
+}
