+++ src/modules/password-reset/presentation/components/password-reset-form-provider/password-reset-form-provider.component.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react'
 import { zodResolver } from '@hookform/resolvers/zod'
 import type { ReactNode } from 'react'
 import { PasswordResetFormSchema } from '../../schemas/password-reset-form.schema'
-import type { PasswordResetFormInterface } from '@/modules/password-reset/domain/interfaces/password-reset.interface'
+import type { PasswordResetFormInterface } from '@/modules/password-reset/domain/interfaces/password-reset-form.interface'
 
 interface PasswordResetFormContextProviderComponentProps {
   children: ReactNode
